fix: deduplicate nested serialization tests
We can deduplicate the nested serialization tests by using
pytest.mark.parametrize to run the same tests with both the default
and pydantic base models. Because the default base model requires the
class to be decorated with @dataclasses.dataclass and the pydantic
base model doesnt, we need to parametrize the decorator as well, and
pass a no-op decorator to the pydantic test.